test auto merge build feature

+++ src/app.py
@@ -1,4 +1,4 @@
-def basic_func(x:int) -> int:
+def basic_func(x:int) -> float:
     try:
         result = 2 / x
         return result
